requirements.txt
dictor
requests
carla-birdeye-view==1.1.1
#opencv-python==4.5.1.48
opencv-python
pygame
tabulate
py-trees==0.8.3
networkx==2.2
Shapely==1.6.4.post2
psutil
xmlschema==1.0.18
ephem
numpy
matplotlib
six
setuptools==49.6.0
torch==1.13.1
torchvision==0.14.1
torchaudio
omegaconf
gym
h5py
hydra-core
timm
stable_baselines3[extra]
pytorch-lightning==1.9.5
imgaug
tqdm
lxml
wandb
ujson
mmcv-full==1.5.3
mmdet==2.25.0
#torch-scatter
diskcache
einops
typing_extensions
jsbeautifier
python-dotenv
#langchain[llms]
timeout-decorator
importlib-metadata==4.13.0
inputimeout
